requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
Pillow==10.4.0
WeasyPrint==62.3
openpyxl==3.1.5
django-cors-headers==4.4.0
python-decouple==3.8
django-extensions==3.2.3
celery==5.3.6
redis==5.0.8
django-redis==5.4.0
whitenoise==6.7.0
gunicorn==22.0.0
dj-database-url==2.1.0
